% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{Constraint}
\alias{Constraint}
\title{Define a set of linearly independent constraints}
\usage{
Constraint(
  A = NULL,
  B = NULL,
  C = NULL,
  a = NULL,
  b = NULL,
  c0 = NULL,
  c1 = NULL,
  n = NULL,
  a_default = 0,
  b_default = 0
)
}
\arguments{
\item{A, a}{A \eqn{k \times n} matrix \code{A} and a \eqn{n} vector \code{a}, defining the equality
constraint \eqn{A x = a}}

\item{B, b}{A \eqn{k \times n} matrix \code{B} and a \eqn{n} vector \code{b}, defining the inequality
constraint \eqn{B x > b}}

\item{C, c0, c1}{A \eqn{k \times n} matrix \code{C} and \eqn{n} vectors \code{c0} and \code{c1}, defining the interval
constraint \eqn{C x \in (c_0,c_1)}}

\item{n}{An integer greater than 0. The dimension of the parameter. Only needed when \code{A},\code{B}, and \code{C} are undefined (thus, only for empty constraints)}

\item{a_default, b_default}{Real numbers. If \code{a} or \code{b} are left undefined, they default to vectors with \code{n} repetitions of \code{a_default}/\code{b_default}.}
}
\value{
An object of class \code{Constraint}.
}
\description{
Define a set of equality, inequality, and/or interval constraints for a parameter of dimension \eqn{n}, using Matrix equation syntax.
}
\details{
The \code{Constraint} class is used to define sets of linearly independent equality, inequality, and interval
constraints. \code{Constraint} objects are created by this function, and by all flavors of helper functions
(e.g., \code{EqualityConstraint}, \code{LinearConstraint}, etc.). Constraints can be combined/intersected with the
\code{+} operator, and this is the suggested way to build up constraints.

\code{Constraint()} acts as the constructor for the \code{Constraint} class. Directly using it
is the most flexible, but also the least "user-friendly" way to specify constraints
(e.g. the exact matrices and vectors need to be figured out beforehand).
}
\examples{
# NULL

}
