% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{EmptyConstraint}
\alias{EmptyConstraint}
\title{Define an empty constraint}
\usage{
EmptyConstraint(n)
}
\arguments{
\item{n}{An integer. The dimension of the parameter}
}
\value{
An object of class \code{Constraint}
}
\description{
Define an empty constraint of dimension \eqn{n}, i.e., a constraint that lets the n-dimensional parameter vary freely.
}
\details{
This is useful for sampling from unconstrained encompassing models, and as a basis for building up more complex constraints
}
\examples{
# NULL

}
