% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fit.SICSModel}
\alias{fit.SICSModel}
\title{Fits a SICS Model (SICSFrame + Constraint)}
\usage{
\method{fit}{SICSModel}(
  Modelobj,
  fit_prior = TRUE,
  fit_posterior = TRUE,
  iter_posterior = 2000,
  chains_posterior = 4,
  iter_prior = 5 * iter_posterior,
  chains_prior = 4,
  force_prior_sampling = FALSE,
  stanmod = Modelobj$stanmod,
  ...
)
}
\description{
In most cases, you will NOT need to manually call this function,
as it is called internally by all model.SICSFrame_* methods.
}
