% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{LinearConstraint}
\alias{LinearConstraint}
\title{... Cannot handle interval constraints}
\usage{
LinearConstraint(equation_string, vars = NULL, n = NULL, default_var = "x")
}
\description{
... Cannot handle interval constraints
}
