% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{model.SICSFrame}
\alias{model.SICSFrame}
\title{Prepare and fit a SICS model (SICSFrame + Constraint)}
\usage{
\method{model}{SICSFrame}(
  SF,
  constraint = NULL,
  fit_prior = TRUE,
  fit_posterior = TRUE,
  no_fit = FALSE,
  no_params = FALSE,
  stanmod = stop("model.SICSFrame called without specifying stanmod"),
  ...
)
}
\description{
This method is called internally by all submethods (model.SICSFrame_*).
}
\details{
Use case for manual use: Development of new SICS Models without overhead.
}
