% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{rbind_padded}
\alias{rbind_padded}
\title{Row-bind padded with zeros}
\usage{
rbind_padded(x, y)
}
\arguments{
\item{x, y}{Two vectors or matrices}
}
\value{
A matrix
}
\description{
A version of \code{rbind} that pads the shorter of the two vectors or matrices with zeroes.
}
\examples{
# NULL

}
