% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sics-package.R
\docType{package}
\name{sics-package}
\alias{sics-package}
\alias{sics}
\title{The 'sics' package.}
\description{
Fit models under linear constraints, using Sampling-In-Constrained-Space (SICS).
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: Lukas Lengersdorff \email{lukaslengersdorff@gmail.com} (\href{https://orcid.org/0000-0002-8750-5057}{ORCID})

}
