% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{SICSFrame_Multinomial}
\alias{SICSFrame_Multinomial}
\title{SICS Frame - Multinomial model}
\usage{
SICSFrame_Multinomial(y, alpha = NULL, alpha_default = 1)
}
\arguments{
\item{y}{Vector of counts}

\item{alpha}{...}

\item{alpha_default}{...}
}
\value{
A SICSFrame_Multinomial object
}
\description{
Collect data and prior information for a multinomial model
}
